Add files via upload
Please consider these for calculator assignment. Thank you.

+++ CalculatorAssignment.java
@@ -14,7 +14,7 @@ public class CalculatorAssignment {
 		return num1*num2;
 	}
 
-	public float div(int num1, int num2) {
+	public float div(float num1, float num2) {
 		return num1/num2;
 	}
